similarity-oracle/requirements.txt
fastapi>=0.115.0
uvicorn==0.34.2
imagehash==4.2.1
pyacoustid==1.3.0
Pillow==12.2.0
numpy>=1.26.0
psycopg2-binary==2.9.10
python-multipart==0.0.20
